+++ src/main/java/at/fhtw/swen/mctg/core/cards/monsters/Knight.java
@@ -3,7 +3,7 @@ package at.fhtw.swen.mctg.core.cards.monsters;
 import at.fhtw.swen.mctg.core.cards.Monster;
 
 public class Knight extends Monster {
-    public Knight(int damage, Element type) {
+    public Knight(double damage, Element type) {
         super(damage, type);
     }
