+++ src/main/java/at/fhtw/swen/mctg/core/cards/monsters/Kraken.java
@@ -3,7 +3,7 @@ package at.fhtw.swen.mctg.core.cards.monsters;
 import at.fhtw.swen.mctg.core.cards.Monster;
 
 public class Kraken extends Monster {
-    public Kraken (int damage, Element type) {
+    public Kraken (double damage, Element type) {
         super(damage, type);
     }
