+++ src/main/java/at/fhtw/swen/mctg/core/cards/monsters/Wizard.java
@@ -3,7 +3,7 @@ package at.fhtw.swen.mctg.core.cards.monsters;
 import at.fhtw.swen.mctg.core.cards.Monster;
 
 public class Wizard extends Monster {
-    public Wizard(int damage, Element type) {
+    public Wizard(double damage, Element type) {
         super(damage, type);
     }
